add precondition validation to build
If a package requires particular users or groups to exist, or
specific system directories to be set up with specific
permissions to work, the options file can now have an
initial_checkup function that verifies them.

Also: fix a couple of stupid typos in the build script; and
modify add_package_user so it removes the mail spool file for
package users (who do not need one).